+++ Models/DTO/Product/ProductPriceRequestDto.cs
@@ -4,6 +4,6 @@ namespace GraduationThesis_CarServices.Models.DTO.Product
     public class ProductPriceRequestDto
     {
         public int ProductId { get; set; }
-        public float ProductPrice { get; set; }
+        public double ProductPrice { get; set; }
     }
 }
